refactor(FormGroup): tvをvariantsの有無で分割しstyle.tsに切り出す
classNameGeneratorのvariantsはfieldsetWithDefaultMarginのみで、これは
Fieldset固有の条件だった。variantsを持たない定義をstyle.tsへ移し、
FormControlはそのまま、Fieldsetはextendして余白調整のvariantを
追加する形にした。

FormGroupはclassNamesをpropsで受け取るだけになり、tvの呼び出しと
fieldsetWithDefaultMargin propsが不要になった。classNameは
classNames.wrapperにマージ済みのためPropsからOmitしている。

variant名はFieldset内では自明なfieldsetを落としてwithDefaultMarginとした。

生成されるクラスが変化しないことは、FormControl・Fieldsetそれぞれの
innerMargin未指定・指定・falsyな境界値である0、className指定、
disabled時の子孫セレクタを含む10パターンの実測比較で確認済み（完全一致）。
