+++ packages/smarthr-ui/src/components/FormGroup/style.ts
@@ -1,0 +1,17 @@
+import { tv } from 'tailwind-variants'
+
+export const classNameGenerator = tv({
+  slots: {
+    wrapper: [
+      'smarthr-ui-FormControl',
+      'shr-mx-[unset] shr-border-none shr-p-[unset]',
+      'disabled:shr-text-disabled',
+      '[&:disabled_.smarthr-ui-FormControl-label_>_span]:shr-text-disabled',
+      '[&:disabled_.smarthr-ui-FormControl-exampleMessage]:shr-text-color-inherit',
+      '[&:disabled_.smarthr-ui-FormControl-errorMessage-Icon]:shr-text-color-inherit',
+      '[&:disabled_.smarthr-ui-FormControl-supplementaryMessage]:shr-text-color-inherit',
+      '[&:disabled_.smarthr-ui-Input]:shr-border-default/50 [&:disabled_.smarthr-ui-Input]:shr-bg-white-darken',
+    ],
+    childrenWrapper: 'smarthr-ui-FormControl-childrenWrapper',
+  },
+})
